Removed redundant files
Correctly leveraging antlr parser to simplify queries
Removed unecessary unit tests due to limited supported types
